requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyodbc==5.0.1
python-dotenv==1.0.0
pydantic==2.5.0
starlette==0.27.0
